Replace preview tree with navigator tree
